fix(backend): fix Pydantic resolution and test parity
- Decouple CIR schemas into dedicated cir_schemas module to solve initialization issues

- Fix SQLite test parity by adding nas_cidr_ranges view in conftest.py

- Increase test coverage to 71% with comprehensive resolution logic testing
